terraform/modules/cloudlab/cloud_functions/increase-budget/requirements.txt
GitPython==3.1.31
requests
pyyaml
